requirements.txt
Django==1.11.3
django-admin-view-permission==1.1
django-flat-responsive==1.4.1
django-stdimage==2.4.1
olefile==0.44
Pillow==4.2.1
progressbar2==3.30.2
psycopg2==2.7.1
python-utils==2.1.0
pytz==2017.2
six==1.10.0
